Refactor motor operations into separate module
